refactor: move context to vm mod
