+++ crates/starlight/src/jsrt/error.rs
@@ -1,7 +1,7 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-use crate::{vm::{Context, arguments::Arguments, error::JsTypeError, error::*, object::JsObject, slot::*, string::JsString, symbol_table::*, value::JsValue}};
+use crate::{vm::{context::Context, arguments::Arguments, error::JsTypeError, error::*, object::JsObject, slot::*, string::JsString, symbol_table::*, value::JsValue}};
 
 pub fn error_constructor(ctx: &mut Context, args: &Arguments) -> Result<JsValue, JsValue> {
     let message = args.at(0).to_string(ctx)?;
