+++ crates/starlight/src/jsrt/function.rs
@@ -1,7 +1,7 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-use crate::{bytecompiler::*, letroot, vm::{Context}, vm::{
+use crate::{bytecompiler::*, letroot, vm::{context::Context}, vm::{
         arguments::Arguments,
         array_storage::ArrayStorage,
         error::JsTypeError,
