+++ crates/starlight/src/jsrt/string.rs
@@ -1,6 +1,6 @@
 use regress::Regex;
 
-use crate::{gc::cell::GcPointer, vm::{Context, arguments::Arguments, array::JsArray, attributes::*, error::{JsRangeError, JsTypeError}, function::JsNativeFunction, object::JsObject, property_descriptor::DataDescriptor, string::{JsString, JsStringObject}, structure::Structure, symbol_table::{Internable, Symbol}, value::*}};
+use crate::{gc::cell::GcPointer, vm::{context::Context, arguments::Arguments, array::JsArray, attributes::*, error::{JsRangeError, JsTypeError}, function::JsNativeFunction, object::JsObject, property_descriptor::DataDescriptor, string::{JsString, JsStringObject}, structure::Structure, symbol_table::{Internable, Symbol}, value::*}};
 use std::{
     char::{decode_utf16, from_u32},
     cmp::{max, min},
