+++ crates/starlight/src/vm/map.rs
@@ -1,4 +1,4 @@
-use super::Context;
+use super::context::Context;
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
