+++ crates/starlight/src/vm/number.rs
@@ -4,7 +4,7 @@
 use crate::prelude::*;
 use std::mem::{size_of, ManuallyDrop};
 
-use super::Context;
+use super::context::Context;
 pub struct NumberObject {
     value: f64,
 }
